requirements.txt
psycopg2==2.9.2
Pillow==8.4.0
pytesseract==0.3.8
requests==2.26.0
selenium==4.0.0